+++ CHANGELOG.md
@@ -8,8 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 - Fixed chats list not updating its order when new messages come in
-- Fixed chat messages sometimes getting lost when you receive a message at the same time 
+- Fixed chat messages sometimes getting lost when you receive a message at the same time
 
+### Added
+- Import/export a muted users
 
 ## [2.1.1] - 2020-09-08
 ### Changed